# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: The Bialto by Asemont Estate API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getAdminMe
      tags: [auth]
      summary: Get current admin session
      responses:
        "200":
          description: Admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated

  # Dashboard
...
  /floors:
    get:
      operationId: getFloors
      tags: [floors]
      summary: List all floors
      responses:
        "200":
          description: List of floors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Floor"
    post:
      operationId: createFloor
      tags: [floors]
      summary: Create a floor (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FloorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Floor"
  /floors/{id}:
    get:
      operationId: getFloor
      tags: [floors]
      summary: Get a floor with its rooms
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Floor detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Floor"
    patch:
      operationId: updateFloor
      tags: [floors]
      summary: Update a floor (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FloorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Floor"
    delete:
      operationId: deleteFloor
      tags: [floors]
      summary: Delete a floor (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Rooms
...
  /bookings:
    get:
      operationId: getBookings
      tags: [bookings]
      summary: List all bookings (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # Attractions
  /attractions:
    get:
      operationId: getAttractions
      tags: [attractions]
      summary: List attractions
      responses:
        "200":
          description: Attractions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attraction"
    post:
      operationId: createAttraction
      tags: [attractions]
      summary: Create attraction (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttractionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attraction"
  /attractions/{id}:
    patch:
      operationId: updateAttraction
      tags: [attractions]
      summary: Update attraction (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttractionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attraction"
    delete:
      operationId: deleteAttraction
      tags: [attractions]
      summary: Delete attraction (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Reviews
  /reviews:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: List approved reviews (public) or all (admin with query)
      parameters:
        - name: approved
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Approve/reject review (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete review (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Gallery
  /gallery:
    get:
      operationId: getGallery
      tags: [gallery]
      summary: List gallery images
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: createGalleryImage
      tags: [gallery]
      summary: Add gallery image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"
  /gallery/{id}:
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete gallery image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Amenities
  /amenities:
    get:
      operationId: getAmenities
      tags: [amenities]
      summary: List amenities
      responses:
        "200":
          description: Amenities list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Amenity"
    post:
      operationId: createAmenity
      tags: [amenities]
      summary: Create amenity (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmenityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Amenity"
  /amenities/{id}:
    patch:
      operationId: updateAmenity
      tags: [amenities]
      summary: Update amenity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmenityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Amenity"
    delete:
      operationId: deleteAmenity
      tags: [amenities]
      summary: Delete amenity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Settings
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, admin]
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/AdminUser"
    AdminUser:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
...
    Floor:
      type: object
      required: [id, name, floorNumber, description]
      properties:
        id:
          type: integer
        name:
          type: string
        floorNumber:
          type: integer
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
    FloorInput:
      type: object
      required: [name, floorNumber, description]
      properties:
        name:
          type: string
        floorNumber:
          type: integer
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
    FloorUpdate:
      type: object
      properties:
        name:
          type: string
        floorNumber:
          type: integer
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
    Room:
      type: object
      required: [id, name, floorId, roomType, pricePerNight, capacity, description, isAvailable, isPublished]
      properties:
        id:
          type: integer
        name:
          type: string
        floorId:
          type: integer
        floorName:
          type: ["string", "null"]
        roomType:
          type: string
        featureImageUrl:
          type: ["string", "null"]
        galleryImages:
          type: array
          items:
            type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        amenities:
          type: array
          items:
            type: string
        description:
          type: string
        isAvailable:
          type: boolean
        isPublished:
          type: boolean
        isFeatured:
          type: boolean
...
    Booking:
      type: object
      required: [id, guestName, guestEmail, guestPhone, checkIn, checkOut, guests, roomType, status, totalAmount, createdAt]
      properties:
        id:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        guests:
          type: integer
        roomId:
          type: ["integer", "null"]
        roomType:
          type: string
        specialRequests:
          type: ["string", "null"]
        status:
          type: string
        totalAmount:
          type: number
        createdAt:
          type: string
    BookingInput:
      type: object
      required: [guestName, guestEmail, guestPhone, checkIn, checkOut, guests, roomType]
      properties:
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        guests:
          type: integer
        roomId:
          type: ["integer", "null"]
        roomType:
          type: string
        specialRequests:
          type: ["string", "null"]
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
        roomId:
          type: ["integer", "null"]
    Attraction:
      type: object
      required: [id, name, description, distance, travelTime]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        distance:
          type: string
        travelTime:
          type: string
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
    AttractionInput:
      type: object
      required: [name, description, distance, travelTime]
      properties:
        name:
          type: string
        description:
          type: string
        distance:
          type: string
        travelTime:
          type: string
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
    AttractionUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        distance:
          type: string
        travelTime:
          type: string
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
    Review:
      type: object
      required: [id, guestName, rating, reviewText, isApproved, createdAt]
      properties:
        id:
          type: integer
        guestName:
          type: string
        guestLocation:
          type: ["string", "null"]
        guestPhoto:
          type: ["string", "null"]
        rating:
          type: integer
        reviewText:
          type: string
        isApproved:
          type: boolean
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [guestName, rating, reviewText]
      properties:
        guestName:
          type: string
        guestLocation:
          type: ["string", "null"]
        guestPhoto:
          type: ["string", "null"]
        rating:
          type: integer
        reviewText:
          type: string
    ReviewUpdate:
      type: object
      properties:
        isApproved:
          type: boolean
    GalleryImage:
      type: object
      required: [id, imageUrl, category, title]
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        category:
          type: string
        title:
          type: string
        sortOrder:
          type: integer
    GalleryImageInput:
      type: object
      required: [imageUrl, category, title]
      properties:
        imageUrl:
          type: string
        category:
          type: string
        title:
          type: string
        sortOrder:
          type: integer
    Amenity:
      type: object
      required: [id, name, icon, description]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
    AmenityInput:
      type: object
      required: [name, icon, description]
      properties:
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
    AmenityUpdate:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
...