# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Bialto by Asemont Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  guestName: string;
  guestEmail: string;
  guestPhone: string;
  checkIn: string;
  checkOut: string;
  guests: number;
  /** @nullable */
  roomId?: number | null;
  roomType: string;
  /** @nullable */
  specialRequests?: string | null;
  status: string;
  totalAmount: number;
  createdAt: string;
}