# lib/api-zod/src/generated/types/bookingUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Bialto by Asemont Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingUpdate {
  status?: string;
  /** @nullable */
  roomId?: number | null;
}