# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Bialto by Asemont Estate API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  name: string;
  floorId: number;
  /** @nullable */
  floorName?: string | null;
  roomType: string;
  /** @nullable */
  featureImageUrl?: string | null;
  galleryImages?: string[];
  pricePerNight: number;
  capacity: number;
  amenities?: string[];
  description: string;
  isAvailable: boolean;
  isPublished: boolean;
  isFeatured?: boolean;
}